🔨 Revamp internal DB working to use custom IDs
So I basically figured out how to move the complicated and crusty seperate-id system to a nicer, custom ID solution.

I also wrote a giant comment on how I'm going to start approaching the API side of things.

I'm also going to try to merge DBexternals.go and DBinterface.go

The difference between the internal DB funcs and the external ones is their types: One returns a prisma object, and the other one returns a struct.
The prisma object can yield more info about the object being retrieved (more info about the author, etc), but the struct is set in stone.

I might even write a bunch of funcs that would theoretically make the two equal. Basically you have a struct, and you call a function to get a specific field on that struct, like the author's info.

TODO:
- ✨ Polish DB interface
- 🔨 Implement the endpoints that dont need auth
- 🔑 Implement some auth method
- 🔑 Integrate auth into endpoints
- ⚠ Error handling
- 😳 Possible speed improvements
- :trollface: 🤏 We do a little testing
- 🛡 Security checks
- 🔨 Middleware