✨ Optimize finding a dweet internally
Not much, just  moved some stuff around, maybe a little bit of a
performance improvement

I keep thinking about how fast my API needs to be,
because I'm stacking a fuck ton of function calls, and I really need to
see which part is the slowest.

Might do a :trollface: bit of testing and profiling, and then I'll see
what parts to Optimize

This is especially useful while making my API, because on one hand I can
call a bunch of functions to give me parts of an object that I can
assemble, or I can call a single function to give me exactly what I
want, but then I'll have to write a bunch of functions, and... yeah. Bad
design.

I want to take the fastest method, it's also probably going to have the
nicest implementation.

TODO:
- ✨ Polish DB internals
- 🔨 Implement the endpoints that dont need auth
- 🔑 Implement some auth method
- 🔑 Integrate auth into endpoints
- ⚠ Error handling
- 😳 Possible speed improvements
- :trollface: 🤏 We do a little testing
- 🛡 Security checks
- 🔨 Middleware