🔨 lastName optional, format schema, sign-up endpoint with hashing
Lots of work done today, just laid the foundations of auth.

For some reason bcrypt returns a new hash even if I run it on the same
params, don't know if that's gonna be a problem later.

I'll keep the testing for later, I'll just do auth for now.

API testing is so much easier now with GraphiQL and the introspection
working.

Played a bunch of Genshin too. Pretty productive day so far.

TODO:
- 🔑 Implement some auth method
- 🔑 Integrate auth into endpoints
- ⚠ Error handling
- 😳 Possible speed improvements
- :trollface: We do a little testing
- 🛡 Security checks
- 🔨 Middleware