🔨 Fix a bunch of bugs, ensure field completeness for some endpoints
Coding when you're tired is not a good idea. Copy pasting code when half asleep will cause bugs.

Continuing on my path to full API completeness.

It has made my API a bit slower, because of all the sub-cases, and the extra queries, but it's what makes the API complete in the first place, so I gotta do it.

Can't make a lightning fast API that doesn't do the job now, can you?

I've implemented all teh endpoints I need for dwitter, the app, and now I'm going to expand on that API to other hypothetical use cases.

TODO:
- 🔎 Check for field presence at each endpoint
- ⚠ Error handling
- 😳 Possible speed improvements
- :trollface: We do a little testing
- 🛡 Security checks
- 🔨 Middleware