🔨 Finish field checking, need to do recursive checks
I've defined a BasicDweet with an author and some other relations, so everytime I return a dweet anywhere I need to make sure to fetch those too.

Right now the biggest problems I'm facing is probably the schema I've created.

The redweet logic is all fucked, and I don't support infinite recursive queries. Basically, the queries I support have a defined maximum depth, because of the whole BasicItem thing, and I think if I want to get rid of this problem, I'll have to use the resolve functions on my schema fields.

Lots of work still left on backend. I thought I was getting close to finishing it, but I think I've messed up a lot of implementation. It's complex when it doesn't need to be.

TODO:
- 🚚 Move to the new redweet logic
- ✨ Get rid of inherent query depth limit
- ⚠ Error handling
- 😳 Possible speed improvements
- :trollface: We do a little testing
- 🛡 Security checks
- 🔨 Middleware