📁 Add CDN, dotenv, connect to firebase.
Nice try, env scrapers. Not today.

I'm very happy with how this turned out. Backend is probably going to be the most satisfying part about this.

Frontend will be a nightmare.

I have an exam in a few days, so I'll have to lay the project off for a while, but I'll def finish the backend before it.

Frontend dev will be a lot slower.

TODO:
- 🛡 Security checks
- 🅱 Naming conventions