Overhaul backend code and fix a shit ton of broken stuff
Holy shit I never knew working on a project for so long rots my brain this much. I made so many mistakes when I was writing my backend.

It feels like I was trying to avoid change by any means possible, and just going with what works, and not what has the most utility.

This commit took me 2 days of quitting everything else in my life and purely writing code. The last 2 days have been the most exhausting.

I'm gonna go play genshin and sleep early today. Maybe I shouldn't be up at 3AM writing a tree of nested if statements for every decision the API needs to make.

The worst part of this commit was that the code wasn't even fixed to work with the different way we handled redweets.

Everything was so broken and I expected the frontend devs to somehow make sense of the broken shit the API spit out.

This codebase was a nightmare, and now it's half a nightmare. It still needs a fuckton of work, but I have no idea how to fix this half-nightmare.

Frontend work is going smoothly though, which I'm happy about.